Finished the swipe to delete functionality, added search functionality on homepage, redesigned some more UI

+++ InfinityNote/Notes/NoteBookController.swift
@@ -10,10 +10,11 @@ import UIKit
 import Firebase
 import Lottie
 
-class NoteBookController: UICollectionViewController, UICollectionViewDelegateFlowLayout {
+class NoteBookController: UICollectionViewController, UICollectionViewDelegateFlowLayout, UISearchBarDelegate {
     
     let cellId = "cellId"
     let headerId = "headerId"
+    let ref = Database.database().reference()
     var animationView: LOTAnimationView = {
         let view = LOTAnimationView(name: "infinityLoader")
         view.contentMode = .scaleAspectFit
@@ -22,6 +23,14 @@ class NoteBookController: UICollectionViewController, UICollectionViewDelegateFl
         return view
     }()
 
+    lazy var searchBar: UISearchBar = {
+        let searchbar = UISearchBar()
+        searchbar.placeholder = "Search for Notebooks"
+        searchbar.endEditing(true)
+        searchbar.delegate = self
+        return searchbar
+    }()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         collectionView?.backgroundColor = paletteSystemWhite
@@ -30,16 +39,26 @@ class NoteBookController: UICollectionViewController, UICollectionViewDelegateFl
         collectionView?.alwaysBounceVertical = true
         
         collectionView?.addSubview(animationView)
-        animationView.frame = CGRect(x: view.center.x/4, y: view.center.y/3, width: 300, height: 300)
+        navigationController?.navigationBar.addSubview(searchBar)
+
         fetchNotebooks()
+        animationView.frame = CGRect(x: view.center.x/4, y: view.center.y/3, width: 300, height: 300)
         setupNavigationController()
     }
     
+    override func viewWillAppear(_ animated: Bool) {
+        searchBar.isHidden = false
+        print("Fetching 1")
+    }
+    
+
+    
     var notebooks = [Notebook]()
+    var filteredNotebooks = [Notebook]()
     func fetchNotebooks(){
         guard let uid = Auth.auth().currentUser?.uid else { return }
         
-        Database.database().reference().child(uid).child("notebooks").observeSingleEvent(of: .value) { (snapshot) in
+        self.ref.child(uid).child("notebooks").observeSingleEvent(of: .value) { (snapshot) in
             guard let dictionaries = snapshot.value as? [String: Any] else { return }
             
             dictionaries.forEach({ (key,value) in
@@ -48,26 +67,44 @@ class NoteBookController: UICollectionViewController, UICollectionViewDelegateFl
                 //self.notebooks.append(notebook)
                 self.notebooks.insert(notebook, at: 0)
             })
-            
-            self.animationView.loopAnimation = false
-            self.animationView.removeFromSuperview()
+            self.filteredNotebooks = self.notebooks
+            print("Fetching 2")
             self.collectionView?.reloadData()
+            
         }
+        self.animationView.loopAnimation = false
+        self.animationView.removeFromSuperview()
+    }
+    
+    // For Searchbar control
+    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
+        if searchText.isEmpty {
+            filteredNotebooks = notebooks
+        }
+        else {
+            filteredNotebooks = self.notebooks.filter({ (notebooks) -> Bool in
+                return notebooks.notebookTitle.lowercased().contains(searchText.lowercased())
+            })
+        }
+        self.collectionView?.reloadData()
     }
     
     func setupNavigationController() {
-        navigationItem.title = "Infinity"
-
-        let image = UIImage(named: "plusIcon")?.withRenderingMode(.alwaysOriginal)
-        navigationItem.rightBarButtonItem = UIBarButtonItem(image: image, style: .plain, target: self, action: #selector(handleAddNotebookButton))
+        //navigationItem.title = "Infinity"
+        let navbar = navigationController?.navigationBar
+        searchBar.anchor(topAnchor: navbar?.topAnchor, bottomAnchor: navbar?.bottomAnchor, leadingAnchor: navbar?.leadingAnchor, trailingAnchor: navbar?.trailingAnchor, paddingTop: 0, paddingBottom: 0, paddingLeft: 8, paddingRight: 8, width: 0, height: 0)
+        collectionView?.keyboardDismissMode = .onDrag
+        
     }
     
     // Need this function in order to apend and insert notebook into collectionView
     func addNotebook(notebook: Notebook) {
         // 1 - modify your array
+        filteredNotebooks.append(notebook)
         notebooks.append(notebook)
         // 2 - insert a new index path into your collecionView
-        let newIndexPath = IndexPath(item: notebooks.count-1, section: 0)
+        let newIndexPath = IndexPath(item: filteredNotebooks.count-1, section: 0)
+        
         self.collectionView?.insertItems(at: [newIndexPath])
     }
     
@@ -86,6 +123,7 @@ class NoteBookController: UICollectionViewController, UICollectionViewDelegateFl
     // referenceSizeForHeaderInSection
     override func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
         let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: headerId, for: indexPath) as! NoteBookHeaderCell
+        
         return header
     }
     
@@ -100,12 +138,12 @@ class NoteBookController: UICollectionViewController, UICollectionViewDelegateFl
     // sizeForItemAt: Requires an extension of UICollectionViewDelegateFlowLayout
     // didSelectItemAt: Checks to see if you clicked on a cell
     override func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return notebooks.count
+        return filteredNotebooks.count
     }
     
     override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellId, for: indexPath) as! NoteBookCell
-        cell.notebook = notebooks[indexPath.item]
+        cell.notebook = filteredNotebooks[indexPath.item]
         return cell
     }
     
@@ -129,20 +167,30 @@ class NoteBookController: UICollectionViewController, UICollectionViewDelegateFl
     override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         let noteController = NoteController(collectionViewLayout: UICollectionViewFlowLayout())
         
-        let notebook = self.notebooks[indexPath.item]
+        let notebook = self.filteredNotebooks[indexPath.item]
         
         noteController.notebookTitle = notebook.notebookTitle
         print(notebook)
+        noteController.searchBar = self.searchBar
         
         navigationController?.pushViewController(noteController, animated: true)
     }
     
     
     override func collectionView(_ collectionView: UICollectionView, performAction action: Selector, forItemAt indexPath: IndexPath, withSender sender: Any?) {
-        notebooks.remove(at: indexPath.row)
-        collectionView.deleteItems(at: [indexPath])
-        print("Here and stuff")
+        let notebookpart = self.filteredNotebooks[indexPath.item]
+        guard let uid = Auth.auth().currentUser?.uid else { return }
         
-        // Here is where I will delete the notebook
+        self.ref.child(uid).child("notebooks").child(notebookpart.notebookTitle).removeValue { (err, ref) in
+            if let err = err {
+                print("Oops, looks like something went wrong: ", err)
+            }
+            print("Successfully removed: ", notebookpart)
+            self.filteredNotebooks.remove(at: indexPath.item)
+            collectionView.deleteItems(at: [indexPath])
+            print("Here and stuff")
+        }
+        
+        
     }
 }
